Problems 349 and 350
Problem 349 (intersection of two arrays) whats a list of all elements that appears in both arrays in input with no repetition.
This can be done by converting the two arrays in sets and using the method .intersection().
Problem 350 is the same but in the result there can be duplicates.
In this case, because the elements of the two arrays are pretty mush in the same amount, we sort the arrays and iterate them both simultaneously incrementing the iterator that is corrisponding to the lower number.