Problem 404 and 409nof leet code
404->sum of left leafs of binary tree
Sol->ric arrive at the leafs and sum its values. When called the recursive function, if we call it on left child we pass 1 if called on right we pass 0.
When arrived at the leaf we moltiply the passed value so we get only left values

409->a palindrom word is compose by letters on the fringe that are even in number and a center that can or can't be odd.
we Counter the string and then we iterate the COunter.
We sum at the lenght of the palindorm word the # of partend letters (
v//2*2)
if the lenght of the palindrom word is even and v is odd we have fund the center of the palindorm and add 1 to ans
else we don't add nothing